Added togglable hotbar
