+++ vue/src/plugins/Provider.ts
@@ -1,13 +1,15 @@
 import { Container } from "@/Models/Container";
+import { Hotbar } from "@/Models/Hotbar";
 import { Inventory } from "@/Models/Inventory";
 import { Plugin } from "vue";
 
-const InventoryPlugin: Plugin = {
+const ProviderPlugin: Plugin = {
   // Track key presses for tab, esc
-  install(app, options: { inventory: Inventory, container: Container }) {
+  install(app, options: { inventory: Inventory, container: Container, hotbar: Hotbar }) {
     app.provide("inventory", options.inventory);
     app.provide("container", options.container);
+    app.provide("hotbar", options.hotbar);
   }
 }
 
-export { InventoryPlugin };
+export { ProviderPlugin };
